@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
-        <TargetFramework>net5.0-windows</TargetFramework>
+        <TargetFramework>net6.0-windows</TargetFramework>
         <RollForward>LatestMajor</RollForward>
         <OutputType>WinExe</OutputType>
         <PublishUrl>publish\</PublishUrl>
@@ -36,6 +36,8 @@
         <GenerateDocumentationFile>True</GenerateDocumentationFile>
     </PropertyGroup>
     <ItemGroup>
+      <None Remove="Resources\tabs_left_button.png" />
+      <None Remove="Resources\tabs_right_button.png" />
       <None Remove="Resources\X_Down.png" />
       <None Remove="X.png" />
     </ItemGroup>
@@ -3650,6 +3652,8 @@
         </None>
         <EmbeddedResource Include="Resources\GML.xshd" />
         <EmbeddedResource Include="Resources\VMASM.xshd" />
+        <Resource Include="Resources\tabs_left_button.png" />
+        <Resource Include="Resources\tabs_right_button.png" />
         <Resource Include="Resources\X.png" />
         <Resource Include="Resources\X_Down.png" />
         <None Update="Repackers\ReduceEmbeddedTexturePages.csx">
@@ -3999,10 +4003,6 @@
         <None Update="DemoScripts\TextInputTestWPF.csx">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
             <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-        </None>
-        <None Update="DemoScripts\TestAUMIConnection.csx">
-          <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-          <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
         </None>
         <None Update="TechnicalScripts\ToggleAlignment.csx">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
