@@ -54,11 +54,21 @@ namespace UndertaleModLib
 
     public interface PrePaddedObject
     {
+
+        /// <summary>
+        /// TODO!
+        /// </summary>
+        /// <param name="writer">Where to serialize to.</param>
         void SerializePrePadding(UndertaleWriter writer);
+
+        /// <summary>
+        /// TODO!
+        /// </summary>
+        /// <param name="reader">Where to deserialize from.</param>
         void UnserializePrePadding(UndertaleReader reader);
     }
 
-    public enum ResourceType : int
+    public enum ResourceType
     {
         None = -1,
         Object = 0,
