@@ -6,7 +6,10 @@ using System.IO;
 
 namespace UndertaleModLib.Util
 {
-    // NOTE: Ported over from DogScepter's QOI converter at https://github.com/colinator27/dog-scepter/
+    /// <summary>
+    /// A class that converts to and from the GM-custom QOI format.
+    /// </summary>
+    /// <remarks>Ported over from DogScepter's QOI converter at <see href="https://github.com/colinator27/dog-scepter/"/>.</remarks>
     public static class QoiConverter
     {
         private const byte QOI_INDEX = 0x00;
@@ -21,19 +24,49 @@ namespace UndertaleModLib.Util
         private const byte QOI_MASK_3 = 0xe0;
         private const byte QOI_MASK_4 = 0xf0;
 
-        public unsafe static Bitmap GetImageFromStream(Stream s)
+        /// <summary>
+        /// Creates a <see cref="Bitmap"/> from a <see cref="Stream"/>.
+        /// </summary>
+        /// <param name="s">The stream to create the PNG image from.</param>
+        /// <returns>The QOI image as a PNG.</returns>
+        /// <exception cref="Exception">If there is an invalid QOIF magic header or there was an error with stride width.</exception>
+        public static Bitmap GetImageFromStream(Stream s)
         {
-            byte[] header = new byte[12];
-            s.Read(header, 0, 12);
+            Span<byte> header = stackalloc byte[12];
+            s.Read(header);
+            int length = header[8] + (header[9] << 8) + (header[10] << 16) + (header[11] << 24);
+            byte[] bytes = new byte[12 + length];
+            s.Position -= 12;
+            s.Read(bytes, 0, bytes.Length);
+            return GetImageFromSpan(bytes);
+        }
+
+        /// <summary>
+        /// Creates a <see cref="Bitmap"/> from a <see cref="ReadOnlySpan"/>.
+        /// </summary>
+        /// <param name="bytes">The <see cref="Span"/> to create the PNG image from.</param>
+        /// <returns>The QOI image as a PNG.</returns>
+        /// <exception cref="Exception">If there is an invalid QOIF magic header or there was an error with stride width.</exception>
+        public static Bitmap GetImageFromSpan(ReadOnlySpan<byte> bytes) => GetImageFromSpan(bytes, out _);
+
+        /// <summary>
+        /// Creates a <see cref="Bitmap"/> from a <see cref="ReadOnlySpan"/>.
+        /// </summary>
+        /// <param name="bytes">The <see cref="Span"/> to create the PNG image from.</param>
+        /// <param name="length">The total amount of data read from the <see cref="Span"/>.</param>
+        /// <returns>The QOI image as a PNG.</returns>
+        /// <exception cref="Exception">If there is an invalid QOIF magic header or there was an error with stride width.</exception>
+        public unsafe static Bitmap GetImageFromSpan(ReadOnlySpan<byte> bytes, out int length)
+        {
+            ReadOnlySpan<byte> header = bytes[..12];
             if (header[0] != (byte)'f' || header[1] != (byte)'i' || header[2] != (byte)'o' || header[3] != (byte)'q')
                 throw new Exception("Invalid little-endian QOIF image magic");
 
             int width = header[4] + (header[5] << 8);
             int height = header[6] + (header[7] << 8);
-            int length = header[8] + (header[9] << 8) + (header[10] << 16) + (header[11] << 24);
+            length = header[8] + (header[9] << 8) + (header[10] << 16) + (header[11] << 24);
 
-            byte[] pixelData = new byte[length];
-            s.Read(pixelData, 0, length);
+            ReadOnlySpan<byte> pixelData = bytes.Slice(12, length);
 
             Bitmap bmp = new Bitmap(width, height, PixelFormat.Format32bppArgb);
 
@@ -47,7 +80,7 @@ namespace UndertaleModLib.Util
             int pos = 0;
             int run = 0;
             byte r = 0, g = 0, b = 0, a = 255;
-            byte[] index = new byte[64 * 4];
+            Span<byte> index = stackalloc byte[64 * 4];
             while (bmpPtr < bmpEnd)
             {
                 if (run > 0)
@@ -126,12 +159,30 @@ namespace UndertaleModLib.Util
 
             bmp.UnlockBits(data);
 
+            length += header.Length;
             return bmp;
         }
 
-        public unsafe static byte[] GetArrayFromImage(Bitmap bmp, int padding = 4)
-        {
-            byte[] res = new byte[(bmp.Width * bmp.Height * 4 * 12) + padding]; // default capacity
+        /// <summary>
+        /// Creates a QOI image as a byte array from a <see cref="Bitmap"/>.
+        /// </summary>
+        /// <param name="bmp">The <see cref="Bitmap"/> to create the QOI image from.</param>
+        /// <param name="padding">The amount of bytes of padding that should be used.</param>
+        /// <returns>A QOI Image as a byte array.</returns>
+        /// <exception cref="Exception">If there was an error with stride width.</exception>
+        public static byte[] GetArrayFromImage(Bitmap bmp, int padding = 4) => GetSpanFromImage(bmp, padding).ToArray();
+
+        /// <summary>
+        /// Creates a QOI image as a <see cref="Span"/> from a <see cref="Bitmap"/>.
+        /// </summary>
+        /// <param name="bmp">The <see cref="Bitmap"/> to create the QOI image from.</param>
+        /// <param name="padding">The amount of bytes of padding that should be used.</param>
+        /// <returns>A QOI Image as a byte array.</returns>
+        /// <exception cref="Exception">If there was an error with stride width.</exception>
+        public unsafe static Span<byte> GetSpanFromImage(Bitmap bmp, int padding = 4) {
+            const int maxChunkSize = 5; // according to the QOI spec: https://qoiformat.org/qoi-specification.pdf
+            const int headerSize = 12;
+            byte[] res = new byte[bmp.Width * bmp.Height * maxChunkSize + headerSize + padding]; // default capacity
             // Little-endian QOIF image magic
             res[0] = (byte)'f';
             res[1] = (byte)'i';
@@ -149,11 +200,11 @@ namespace UndertaleModLib.Util
             byte* bmpPtr = (byte*)data.Scan0;
             byte* bmpEnd = bmpPtr + (4 * bmp.Width * bmp.Height);
 
-            int resPos = 12;
+            int resPos = headerSize;
             byte r = 0, g = 0, b = 0, a = 255;
             int run = 0;
             int v = 0, vPrev = 0xff;
-            int[] index = new int[64];
+            Span<int> index = stackalloc int[64];
             while (bmpPtr < bmpEnd)
             {
                 b = *bmpPtr;
@@ -245,13 +296,13 @@ namespace UndertaleModLib.Util
             resPos += padding;
 
             // Write final length
-            int length = resPos - 12;
+            int length = resPos - headerSize;
             res[8] = (byte)(length & 0xff);
             res[9] = (byte)((length >> 8) & 0xff);
             res[10] = (byte)((length >> 16) & 0xff);
             res[11] = (byte)((length >> 24) & 0xff);
 
-            return res[..resPos];
+            return res.AsSpan()[..resPos];
         }
     }
 }
